Disc_dynamics.jl

Functions

ContGridModML.NRsolver — Method

NRsolver(
    Ybus::SparseArrays.SparseMatrixCSC{<:Complex, <:Int64},
    V::Vector{<:Real},
    theta::Vector{<:Real},
    p::Vector{<:Real},
    q::Vector{<:Real},
    idpq::Vector{<:Int64},
    id_slack::Int64;
    tol,
    maxiter
) -> Tuple{Vector{<:Real}, Vector{<:Real}, Int64}

Use the Newton Raphson method to solve the powerflow equations.This method is adapted from its version on the Pantagruel repository (https://doi.org/10.5281/zenodo.2642175). For information on solving the power flow equations with Newton-Raphson, see, for instance, V. Vittal and A. Bergen, Power systems analysis, Prentice Hall, 1999.

ContGridModML.disc_dynamics — Method

disc_dynamics(
    dm::ContGridModML.DiscModel,
    tstart::Real,
    tend::Real,
    delta_p::Union{Real, Vector{<:Real}},
    faultid::Int64;
    dt,
    tol,
    maxiter,
    dmin,
    alg,
    solve_kwargs
) -> SciMLBase.ODESolution