Initl.jl

Functions

ContGridModML.distribute_load! — Method

distribute_load!(
    cm::ContGridModML.ContModel,
    dm::ContGridModML.DiscModel
)

ContGridModML.init_model — Method

init_model(
    grid::Ferrite.Grid
) -> ContGridModML.ContModel{Float64}

Create a continuous model from a discrete model by using a diffusion process to distribute the paramters.

ContGridModML.integrate — Method

integrate(
    dh::Ferrite.DofHandler,
    cellvalues::Ferrite.CellScalarValues,
    q_proj::SparseArrays.SparseMatrixCSC,
    vals::Vector{<:Real}
) -> Real

Integrate a function over the whole area of the grid using the finite element method.

ContGridModML.interpolate — Method

interpolate(
    x::Array{Tensors.Tensor{1, 2, T<:Real, 2}, 1},
    grid::Ferrite.Grid,
    dh::Ferrite.DofHandler,
    u::Vector{<:Real},
    fname::Symbol;
    off,
    factor,
    extrapolate,
    warn
) -> Vector{T} where T<:Real

Interpolate values from the continues model from a vector of coordinates. If a given coordinate is outside the grid it is replaced by the closed value on the grid.

ContGridModML.interpolate — Method

interpolate(
    x::Tensors.Tensor{1, 2, T<:Real, 2},
    grid::Ferrite.Grid,
    dh::Ferrite.DofHandler,
    u::Vector{<:Real},
    fname::Symbol;
    off,
    factor,
    extrapolate,
    warn
) -> Real

Interpolate values from the continues model from a coordinate. If the given coordinate is outside the grid it is replaced by the closed value on the grid.

ContGridModML.set_local_disturbance! — Method

set_local_disturbance!(
    cm::ContGridModML.ContModel,
    coord::Vector{<:Real},
    dP::Real
)

ContGridModML.set_slack! — Method

set_slack!(
    cm::ContGridModML.ContModel,
    dm::ContGridModML.DiscModel
)