Tools.jl

Description of various helper functions. At the moment they are all related to loading and saving the solution structures.

Contents

Contents

Index

Functions

Index

ContGridModML.albers_projection

ContGridModML.cont_from_dict

ContGridModML.cont_to_dict

ContGridModML.dict_to_dynamic

ContGridModML.dict_to_hdf5

ContGridModML.dict_to_sol

ContGridModML.dict_to_sparse

ContGridModML.dict_to_static

ContGridModML.disc_from_dict

ContGridModML.disc_to_dict

ContGridModML.distribute_country_load

ContGridModML.dynamic_to_dict

ContGridModML.hdf5_to_dict

ContGridModML.import_border

ContGridModML.load_discrete_model

ContGridModML.load_discrete_model_from_json

ContGridModML.load_discrete_model_from_powermodels

ContGridModML.load_model

ContGridModML.load_sol

ContGridModML.model_to_dict

ContGridModML.opf_from_country

ContGridModML.remove_nan

ContGridModML.save_model

ContGridModML.save_sol

ContGridModML.sol_to_dict

ContGridModML.sparse_to_dict

ContGridModML.stable_sol!

ContGridModML.static_to_dict

Functions

ContGridModML.albers_projection — Method

albers_projection(
    coord::Matrix{<:Real};
    lon0,
    lat0,
    lat1,
    lat2,
    R
) -> Matrix{<:Real}

Apply the Albers projection to a vector of coordinates. The coordinates need to be given as latitude, longitude. See https://en.wikipedia.org/wiki/Albers_projection

ContGridModML.cont_from_dict — Method

cont_from_dict(
    data::Dict{String}
) -> ContGridModML.ContModel

ContGridModML.cont_to_dict — Method

cont_to_dict(
    cm::ContGridModML.ContModel
) -> Dict{String, Any}

ContGridModML.dict_to_dynamic — Method

dict_to_dynamic(
    data::Dict{String}
) -> ContGridModML.DynamicSol

Load a dynamic solution from a dictionary.

ContGridModML.dict_to_hdf5 — Method

dict_to_hdf5(data::Dict, fid::HDF5.H5DataStore)

ContGridModML.dict_to_sol — Method

dict_to_sol(
    data::Dict{String}
) -> Union{ContGridModML.DynamicSol, ContGridModML.StaticSol}

Load a solution from a dictionary.

ContGridModML.dict_to_sparse — Method

dict_to_sparse(
    dict::Dict{String, Any}
) -> SparseArrays.SparseMatrixCSC

ContGridModML.dict_to_static — Method

dict_to_static(
    data::Dict{String}
) -> ContGridModML.StaticSol

Load a static solution from a dictionary.

ContGridModML.disc_from_dict — Method

disc_from_dict(
    dict::Dict{String}
) -> ContGridModML.DiscModel

ContGridModML.disc_to_dict — Method

disc_to_dict(dm::ContGridModML.DiscModel) -> Dict{String}

ContGridModML.distribute_country_load — Method

distribute_country_load(
    grid::Dict{String, Any},
    country::Dict{String, <:Real}
) -> Dict{String, Any}

ContGridModML.dynamic_to_dict — Method

dynamic_to_dict(
    sol::ContGridModML.DynamicSol
) -> Dict{String, Any}

Turn a dynamic solution into a dictionary.

ContGridModML.hdf5_to_dict — Method

hdf5_to_dict(fid::HDF5.H5DataStore) -> Dict{String, Any}

ContGridModML.import_border — Method

import_border(filename::String) -> Any

Import border from a json file, apply the Albers projection and rescale it such that the longest dimension is 1. The coordinates need to be given as latitude, longitude.

ContGridModML.load_discrete_model — Method

load_discrete_model(
    dataname::String,
    scaling_factor::Float64
) -> ContGridModML.DiscModel

Load a discrete model from a file and rescale the coordinates to match the continuous model.

ContGridModML.load_discrete_model_from_json — Method

load_discrete_model_from_json(
    dataname::String,
    project::Bool,
    scale_factor::Real
) -> ContGridModML.DiscModel

ContGridModML.load_discrete_model_from_powermodels — Method

load_discrete_model_from_powermodels(
    data::Dict{String, Any},
    project::Bool,
    scale_factor::Real
) -> ContGridModML.DiscModel

ContGridModML.load_model — Method

load_model(fn::String) -> Any

Load a continuous or discrete model from a hdf5 file.

ContGridModML.load_sol — Method

load_sol(
    fn::String
) -> Union{ContGridModML.DynamicSol, ContGridModML.StaticSol}

Load a solution from a HDF5 file.

ContGridModML.model_to_dict — Method

model_to_dict(
    model::ContGridModML.GridModel
) -> Dict{String}

ContGridModML.opf_from_country — Method

opf_from_country(
    grid::Dict{String, Any},
    country::Dict{String, <:Real}
) -> Dict{String, Any}

ContGridModML.remove_nan — Method

remove_nan(grid::Dict{String, Any}) -> Dict{String, Any}

ContGridModML.save_model — Method

save_model(fn::String, model::ContGridModML.GridModel)

Save a continuous or discrete model to a hdf5 file.

ContGridModML.save_sol — Method

save_sol(fn::String, sol::ContGridModML.ContSol)

Save a solution to a HDF5 file.

ContGridModML.sol_to_dict — Method

sol_to_dict(sol::ContGridModML.ContSol) -> Dict{String, Any}

Turn a solution into a dict.

ContGridModML.sparse_to_dict — Method

sparse_to_dict(
    sm::SparseArrays.SparseMatrixCSC
) -> Dict{String}

ContGridModML.stable_sol! — Method

stable_sol!(cm::ContGridModML.ContModel)

ContGridModML.static_to_dict — Method

static_to_dict(
    sol::ContGridModML.StaticSol
) -> Dict{String, Any}

Turn a static solution into a dictionary.