Dynamic.jl

Description of internally used functions for the learning of the dynamic parameters.

Contents

Dynamic.jl

Contents

Index

Functions

Index

ContGridModML.assemble_f_dynamic

ContGridModML.assemble_matrices_dynamic

ContGridModML.cont_dyn

ContGridModML.disc_dyn

ContGridModML.gen_idxs

ContGridModML.generate_comp_idxs

ContGridModML.grad

ContGridModML.grad_proj

ContGridModML.init_expansion

ContGridModML.initial_conditions

ContGridModML.lambda_dyn

ContGridModML.lap_eigenvectors

ContGridModML.loss

ContGridModML.projectors_dynamic

ContGridModML.simul

ContGridModML.test_loss

ContGridModML.trapz

ContGridModML.update

Functions

ContGridModML.assemble_f_dynamic — Function

assemble_f_dynamic(
    cm::ContGridModML.ContModel,
    dm::ContGridModML.DiscModel,
    fault_ix::Vector{<:Integer},
    dP::Union{Real, Vector{<:Real}},
    Af::SparseArrays.SparseMatrixCSC,
    q_proj::SparseArrays.SparseMatrixCSC,
    σ::Real
) -> Matrix{Float64}

Assemble all the force vectors for the dynamical simulations.

ContGridModML.assemble_matrices_dynamic — Function

assemble_matrices_dynamic(
    model::ContGridModML.ContModel
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64}, SparseArrays.SparseMatrixCSC{Float64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, Matrix{Float64}}

Create all the necessary finite element matrices from a given model.

The matrices returned are

M_const the constant (non-parameter dependent) part of the mass matrix

K_const the constant (non-parameter dependent) part of the stiffness matrix

A the matrix needed to create the non-constant part of both the mass and stiffness matrices. For example the mass matrix can be calculated as M = M_const + A * m_quad * A'

Af the matrix needed to create the force vector as f = Af * (p_quad + fault_quad)

q_coords Coordinates of all quadrature points in the same order as stored in the DoF-handler

ContGridModML.cont_dyn — Function

cont_dyn(
    M::SparseArrays.SparseMatrixCSC,
    K::SparseArrays.SparseMatrixCSC,
    f::Vector{<:Real},
    u₀::Vector{<:Real},
    tf::Real;
    solve_kwargs
) -> SciMLBase.ODESolution

Run a dynamical simulation of the continuous model.

ContGridModML.disc_dyn — Function

disc_dyn(
    dm::ContGridModML.DiscModel,
    fault_node::Integer,
    fault_size::Real,
    dt::Real,
    tf::Real;
    solve_kwargs
) -> SciMLBase.ODESolution

Run a dynamical simulation of the discrete model.

ContGridModML.gen_idxs — Function

gen_idxs(
    dm::ContGridModML.DiscModel,
    dP::Real,
    n_train::Integer,
    n_test::Int64;
    seed
) -> Tuple{Vector{<:Integer}, Vector{<:Integer}}

Randomly choose generators for the training and test sets.

ContGridModML.generate_comp_idxs — Function

generate_comp_idxs(
    cm::ContGridModML.ContModel,
    dm::ContGridModML.DiscModel,
    tri::Vector{<:Integer},
    tei::Vector{<:Integer},
    n::Int64
) -> Vector{<:Integer}

Generate a list of indices used for the comparison of the ground truth with the results from the continuous model.

An equally spaced grid is overlaid over the area. If the points are within the area, the closest bus in the discrete model is found and added to the list of indices. The generators of the test and training set are not eligible for comparison and are remove from the list of possible indices. The number of points can be controlled by n, which gives the number of points in the largest dimension.

ContGridModML.grad — Function

grad(
    sol_cont::SciMLBase.ODESolution,
    sol_lambda::SciMLBase.ODESolution,
    g_proj::Vector{SparseArrays.SparseMatrixCSC}
) -> Vector{Float64}
grad(
    sol_cont::SciMLBase.ODESolution,
    sol_lambda::SciMLBase.ODESolution,
    g_proj::Vector{SparseArrays.SparseMatrixCSC},
    dt::Real
) -> Vector{Float64}

Calculate the gradient from the solution of the adjoint dynamics.

ContGridModML.grad_proj — Function

grad_proj(
    A::SparseArrays.SparseMatrixCSC,
    q_proj::SparseArrays.SparseMatrixCSC,
    evecs::Matrix{<:Real},
    n_coeffs::Integer
) -> Vector{SparseArrays.SparseMatrixCSC}

Calculate the projection matrix of the discrete solution onto the adjoint solution.

ContGridModML.init_expansion — Function

init_expansion(
    cm::ContGridModML.ContModel,
    eve::Matrix{<:Real},
    n_modes::Integer,
    n_coeffs::Integer
) -> Tuple{Vector{<:Real}, Matrix{<:Real}}

Create the initial values for the parameters.

The parameters are expanded in eigenvectors of the grid Laplacian. The first n_modes eigenvectors are chosen. The first n_coeffs coefficients are chosen by projecting the results of the heat equation diffusion onto the eigenvectors. The n_modes - n_coeffs are set to zero.

ContGridModML.initial_conditions — Function

initial_conditions(
    cm::ContGridModML.ContModel
) -> Vector{Float64}

Create the initial conditions for the continuous simulations.

ContGridModML.lambda_dyn — Function

lambda_dyn(
    cont_sol::SciMLBase.ODESolution,
    disc_sol::SciMLBase.ODESolution,
    M::SparseArrays.SparseMatrixCSC,
    K::SparseArrays.SparseMatrixCSC,
    ω_proj::SparseArrays.SparseMatrixCSC,
    tf::Real,
    idxs::Vector{<:Integer};
    solve_kwargs
) -> SciMLBase.ODESolution

Solve the adjoint dynamics.

The continuous and discrete solutions are needed as well as the comparison indices to calculate the contributions from the loss function.

ContGridModML.lap_eigenvectors — Function

lap_eigenvectors(
    cm::ContGridModML.ContModel
) -> Matrix{Float64}

Calculate the eigenvectors of the unweighted Laplacian of the grid used for the continuous simulations.

ContGridModML.loss — Function

loss(
    sol_cont::SciMLBase.ODESolution,
    sol_disc::SciMLBase.ODESolution,
    ω_proj::SparseArrays.SparseMatrixCSC,
    idxs::Vector{<:Integer}
) -> Vector{Float64}

Calculate the value of the loss function.

ContGridModML.projectors_dynamic — Function

projectors_dynamic(
    cm::ContGridModML.ContModel,
    dm::ContGridModML.DiscModel,
    q_coords::Matrix{<:Real},
    ω_idxs::Vector{<:Integer}
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}

Create projectors for nodal values onto quadrature points and onto comparison locations.

Returned projectors

q_proj Project the nodal values onto the quadrature points

ω_proj Project the nodal values onto the given comparison points. This is used to calculate the loss function.

ContGridModML.simul — Function

simul(
    disc_sol::SciMLBase.ODESolution,
    M_const::SparseArrays.SparseMatrixCSC,
    K_const::SparseArrays.SparseMatrixCSC,
    m::Vector{<:Real},
    d::Vector{<:Real},
    f::Vector{<:Real},
    A::SparseArrays.SparseMatrixCSC,
    q_proj::SparseArrays.SparseMatrixCSC,
    ω_proj::SparseArrays.SparseMatrixCSC,
    g_proj::Vector{SparseArrays.SparseMatrixCSC},
    idxs::Vector{<:Integer},
    u₀::Vector{<:Real},
    tf::Real;
    cont_kwargs,
    lambda_kwargs
) -> Tuple{Vector{Float64}, Float64}

Do a full simulation step for one data set.

The continuous solution is calculated first and then used to obtain the adjoint solution. Afterwards, the gradient and value of the loss function are calculated and returned.

ContGridModML.test_loss — Function

test_loss(
    disc_sols::Vector{SciMLBase.ODESolution},
    M_const::SparseArrays.SparseMatrixCSC,
    K_const::SparseArrays.SparseMatrixCSC,
    m::Vector{<:Real},
    d::Vector{<:Real},
    f_test::Matrix{<:Real},
    A::SparseArrays.SparseMatrixCSC,
    q_proj::SparseArrays.SparseMatrixCSC,
    ω_proj::SparseArrays.SparseMatrixCSC,
    idxs::Vector{<:Integer},
    u₀::Vector{<:Real},
    tf::Real;
    cont_kwargs
) -> Vector{Float64}

Calculate the loss values for the test data sets.

ContGridModML.trapz — Function

trapz(
    t₀::Real,
    tf::Real,
    dt::Real,
    int!::Function,
    N::Integer
) -> Vector{Float64}

Trapezoidal rule for the integration of a given function.

ContGridModML.update — Function

update(
    p::Vector{<:Real},
    g::Vector{<:Real},
    eve::Matrix{<:Real},
    i::Integer,
    f::Function
) -> Vector{<:Real}

Update the parameters using a restricted gradient descent to ensure positiveness.