Static.jl

Description of internally used functions for the learning of the static parameters.

Contents

Static.jl

Contents

Index

Functions

Index

ContGridModML.assemble_disc_theta

ContGridModML.assemble_f_static

ContGridModML.assemble_matrices_static

ContGridModML.check_slack

ContGridModML.discrete_models

ContGridModML.get_losses

ContGridModML.prediction

ContGridModML.projectors_static

ContGridModML.susceptances

Functions

ContGridModML.assemble_disc_theta — Function

assemble_disc_theta(
    training::Vector{ContGridMod.DiscModel},
    test::Vector{ContGridMod.DiscModel}
) -> Tuple{Matrix{Float64}, Matrix{Float64}}

Assemble all the ground truth data into one matrix for the training and one for the test sets.

ContGridModML.assemble_f_static — Function

assemble_f_static(
    model::ContGridMod.ContModel,
    training::Vector{ContGridMod.DiscModel},
    test::Vector{ContGridMod.DiscModel},
    Af::SparseArrays.SparseMatrixCSC,
    q_proj::SparseArrays.SparseMatrixCSC;
    tf,
    κ,
    σ
) -> Tuple{Matrix{Float64}, Matrix{Float64}}

Assemble the force vectors for the static solutions.

ContGridModML.assemble_matrices_static — Function

assemble_matrices_static(
    model::ContGridMod.ContModel
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, Matrix{Float64}}

Create all the necessary finite element matrices from a given model.

The returned matrices are

Af the matrix needed to create the force vector as f = Af * p_quad

Ak the matrix needed to create the stiffness matrix as K = A * b_quad * A'. The susceptances need to be ordered as $b_x(\mathbf{q_1}), b_y(\mathbf{q_1}), b_x(\mathbf{q_2}),\dots$

dim matrix with single entry for the slack bus to ensure well-posedness of the system of linear equations

q_coords Coordinates of the quadrature points in the same order as stored in the DoF-handler

ContGridModML.check_slack — Function

check_slack(
    training::Vector{ContGridMod.DiscModel},
    test::Vector{ContGridMod.DiscModel}
) -> Bool

Check if all the slack buses in the training and test data sets are the same.

ContGridModML.discrete_models — Function

discrete_models(
    train_fn::String,
    test_fn::String,
    n_train::Integer,
    n_test::Integer,
    scale_factor::Real
) -> Tuple{Vector{ContGridMod.DiscModel}, Vector{ContGridMod.DiscModel}}

Load all the discrete models for the training and test data sets.

ContGridModML.get_losses — Function

get_losses(
    train_pred::Matrix{<:Real},
    test_pred::Matrix{<:Real},
    t_train::Matrix{<:Real},
    t_test::Matrix{<:Real};
    δ
) -> Tuple{Vector{<:Real}, Vector{<:Real}}

Obtain the loss values for the training and test data sets.

ContGridModML.prediction — Function

prediction(
    K::SparseArrays.AbstractSparseMatrix,
    f_train::Matrix{<:Real},
    f_test::Matrix{<:Real},
    proj::SparseArrays.AbstractSparseMatrix
) -> Tuple{Matrix{<:Real}, Matrix{<:Real}}

Obtain the prediction of the stable solution for the training and test data sets.

ContGridModML.projectors_static — Function

projectors_static(
    model::ContGridMod.ContModel,
    dm::ContGridMod.DiscModel,
    q_coords::Matrix{<:Real}
) -> Tuple{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseMatrixCSC{Float64, Int64}}

Projectors of nodal values onto the discrete values and the quadrature points.

The returned matrices are

θ_proj project the nodal values onto the discrete nodes for comparison

q_proj project the nodal values onto the quadrature points

q_proj_b project the susceptances onto the quadrature points. The susceptances need to be ordered as $b_x(\mathbf{r_1}), b_y(\mathbf{r_1}), b_x(\mathbf{r_2}),\dots$

ContGridModML.susceptances — Function

susceptances(
    A::SparseArrays.AbstractSparseMatrix,
    dim::SparseArrays.AbstractSparseMatrix,
    q_proj::SparseArrays.AbstractSparseMatrix,
    proj::SparseArrays.AbstractSparseMatrix,
    f_train::Matrix{<:Real},
    t_train::Matrix{<:Real},
    b::Vector{<:Real},
    n_epochs::Integer,
    n_batches::Int64;
    bmin,
    seed,
    δ
) -> Tuple{Vector{<:Real}, Matrix{<:Real}}

Do the actual learning of the parameters.